% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{checkforfactor}
\alias{checkforfactor}
\title{Checking factors}
\usage{
checkforfactor(DF, vars = colnames(DF), verbose = TRUE)
}
\arguments{
\item{DF}{dataframe : the dataframe to clean}

\item{verbose}{(optional) logical : TRUE will display some informations in the console}
}
\description{
Checking factors
}
