% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_explicatives_matrix}
\alias{get_explicatives_matrix}
\title{Explicatives matrix}
\usage{
get_explicatives_matrix(
  explicatives,
  DF = NULL,
  names = explicatives,
  complete_names = TRUE,
  one_lines = TRUE,
  make.names = TRUE
)
}
\arguments{
\item{explicatives}{vector}

\item{names}{vector: names of the explicatives}

\item{complete_names}{vector :}

\item{one_lines}{vector of boolean : whether to display one or two lines for 2 levels factors}

\item{data}{}
}
\value{

}
\description{
Explicatives matrix
}
