% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_y_vector}
\alias{get_y_vector}
\title{Get y vector}
\usage{
get_y_vector(y, DF = NULL, name = y, y_test = y, names = NA)
}
\arguments{
\item{y}{}

\item{DF}{}

\item{name}{}

\item{y_test}{}

\item{names}{}
}
\description{
Get y vector
}
