% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{make_palette}
\alias{make_palette}
\title{Palette}
\usage{
make_palette(
  pal = c("few", "RdBu"),
  n = 8,
  type = c("Medium", "Light", "Dark"),
  show = FALSE
)
}
\arguments{
\item{pal}{(character) : name of the palette}

\item{n}{(numeric) : number of colors to display}

\item{type}{(character) : for few palette, Medium, Light or Dark}

\item{show}{(logical) : wether to show the palette}
}
\value{

}
\description{
Palette
}
