% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{makepalette}
\alias{makepalette}
\title{Makepalette}
\usage{
makepalette(
  palette = few_pal(),
  n = 8,
  type = c("Medium", "Light", "Dark"),
  show = FALSE
)
}
\arguments{
\item{palette}{}

\item{show}{}
}
\value{

}
\description{
Makepalette
}
