% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{palette}
\alias{palette}
\title{Palette}
\usage{
palette(
  palette = c("few", "RdBu"),
  n = 8,
  type = c("Medium", "Light", "Dark"),
  show = FALSE
)
}
\arguments{
\item{palette}{}

\item{show}{}
}
\value{

}
\description{
Palette
}
