% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{sankeyplot}
\alias{sankeyplot}
\title{Sankey plot}
\usage{
sankeyplot(links, colors = F)
}
\arguments{
\item{matrix}{: a dataframe with source, target and value columns}

\item{labels}{}

\item{titles}{}

\item{values}{}
}
\description{
Sankey plot
}
