% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{survivor}
\alias{survivor}
\title{Survivor}
\usage{
survivor(
  DF,
  explicatives,
  time = "time",
  event = "event",
  names = explicatives,
  keep = 1,
  verbose = TRUE,
  round = 2
)
}
\arguments{
\item{DF}{dataframe : data to be analysed}

\item{explicatives}{vecor of character : columns of the dataframe to be analysed}

\item{time}{character : name of the column representing time to event}

\item{event}{character : name of the column representing the event}

\item{names}{vector of character : name of the explicatives variables as it will be display on table}

\item{keep}{vector of character : explicatives that should never be erased of thee model}
}
\value{
dataframe, excel file or html table depending on the exit parameter
}
\description{
Survivor
}
