% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{tphistogram}
\alias{tphistogram}
\title{Histogram}
\usage{
tphistogram(data, x, y, colors, position = "stack", ...)
}
\arguments{
\item{data}{: dataframe}

\item{x}{: column use to plot the histogram}

\item{y}{: column that can be used to separate the data}

\item{colors}{: character or vector of colors}

\item{position}{: identity, stack or dodge}

\item{...}{: other arguments to be passed through gghistogram}
}
\description{
Histogram
}
