% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Table 1}
\usage{
table1(
  DF,
  explicatives,
  y = NULL,
  overall = TRUE,
  tests = "classic",
  strata = NULL,
  adjusted = NULL,
  mutation = 10,
  round = c(3, 2),
  exit = "console",
  num_function = "auto",
  available_data = TRUE,
  by_line = FALSE,
  complete = F
)
}
\arguments{
\item{DF}{dataframe : datas to be analysed}

\item{explicatives}{: columns of the dataframe to be analysed}

\item{y}{character : column that will separate the dataframe}

\item{overall}{booleen (optionnal) : TRUE if an "overall" column is wanted}

\item{tests}{character / vector (optionnal) : classical (student, khi2...) or logistic regression. Specify if paired.}

\item{strata}{: columns with the number of the paires}

\item{mutation}{numeric : number of modalities to display for one variable. If there is more than "mutation" categories, the modalities after this threeshold are wrapped into an "others" categorie.}

\item{round}{(optionnal) : number of digits to be display in the results. If a vector is provided, digits stands for p-value then variables.}

\item{exit}{character : 'html', 'console' or 'excel'}

\item{y_test}{character (optionnal) : name of the column that will be used for statistical test}
}
\value{
dataframe, excel file or html table depending on the exit parameter
}
\description{
Table 1
}
